added icons for TimeFilterToggle

+++ shared/src/commonMain/kotlin/io.github.krisbitney.yuli/repository/BackgroundTaskLauncher.kt
@@ -2,6 +2,7 @@ package io.github.krisbitney.yuli.repository
 
 import io.github.krisbitney.yuli.api.SocialApiFactory
 import io.github.krisbitney.yuli.database.YuliDatabase
+import io.github.krisbitney.yuli.models.Event
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.IO
 import kotlinx.coroutines.withContext
@@ -9,8 +10,6 @@ import kotlinx.datetime.Clock
 import kotlinx.datetime.Instant
 
 const val UPDATE_FOLLOWS_INTERVAL_SECONDS: Long = 60 * 60 * 6 // 6 hours
-
-const val DAYS_TO_KEEP_EVENTS: Int = 30
 
 expect object BackgroundTaskLauncher {
     fun <AndroidContext> updateFollowsAndNotify(context: AndroidContext)
@@ -27,14 +26,16 @@ object BackgroundTasks {
                 val e = Exception("No user logged in")
                 return@use Result.failure(e)
             }
-            // clear old events to keep app storage down
-            val oldEvents = db.selectEvents(
-                Instant.DISTANT_PAST.epochSeconds,
-                db.daysAgoUnixTimestamp(DAYS_TO_KEEP_EVENTS)
-            )
-            db.deleteEvents(oldEvents)
             // update follows
             val result = ApiHandler(api, db).updateFollows(username, reportProgress)
+            if (result.isSuccess) {
+                // clear old events to keep app storage down
+                val oldEvents = db.selectEvents(
+                    Instant.DISTANT_PAST.epochSeconds,
+                    db.daysAgoUnixTimestamp(Event.DAYS_TO_KEEP_EVENTS)
+                )
+                db.deleteEvents(oldEvents)
+            }
             // record update time
             db.selectState()?.let {
                 db.insertOrReplaceState(it.copy(lastUpdate = Clock.System.now().epochSeconds))
